20762: Move reaction class to schemas
