+++ howso/client/schemas/__init__.py
@@ -1,4 +1,5 @@
 from .base import BaseSchema
 from .project import Project, ProjectDict
+from .reaction import Reaction
 from .session import Session, SessionDict
 from .trainee import Trainee, TraineeDict, TraineePersistence
