+++ howso/direct/client.py
@@ -46,7 +46,7 @@ from howso.client.base import AbstractHowsoClient
 from howso.client.cache import TraineeCache
 from howso.client.configuration import HowsoConfiguration
 from howso.client.exceptions import HowsoError, UnsupportedArgumentWarning
-from howso.client.schemas import Session, Trainee
+from howso.client.schemas import Reaction, Session, Trainee
 from howso.engine.typing import Library, Persistence
 from howso.utilities import (
     build_react_series_df,
@@ -59,7 +59,6 @@ from howso.utilities import (
     validate_list_shape
 )
 from howso.utilities.feature_attributes.base import MultiTableFeatureAttributes, SingleTableFeatureAttributes
-from howso.utilities.reaction import Reaction
 from .core import HowsoCore
 
 # Client version
